move HyperionScannerTask to its own class
