+++ mobile/src/main/java/com/abrenoch/hyperiongrabber/mobile/MainActivity.java
@@ -8,7 +8,6 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.graphics.Color;
 import android.media.projection.MediaProjectionManager;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.RequiresApi;
@@ -23,7 +22,6 @@ import android.widget.ImageView;
 import android.widget.Toast;
 
 import com.abrenoch.hyperiongrabber.common.HyperionScreenService;
-import com.abrenoch.hyperiongrabber.common.network.NetworkScanner;
 
 public class MainActivity extends AppCompatActivity implements ImageView.OnClickListener,
         ImageView.OnFocusChangeListener {
@@ -61,8 +59,6 @@ public class MainActivity extends AppCompatActivity implements ImageView.OnClick
         LocalBroadcastManager.getInstance(this).registerReceiver(
                 mMessageReceiver, new IntentFilter(HyperionScreenService.BROADCAST_FILTER));
         checkForInstance();
-
-        new HyperionScannerTask().execute();
     }
 
     @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
@@ -175,38 +171,7 @@ public class MainActivity extends AppCompatActivity implements ImageView.OnClick
         return false;
     }
 
-    private static class HyperionScannerTask extends AsyncTask<Void, Float, String> {
 
-        @Override
-        protected String doInBackground(Void... voids) {
-            Log.d("Hyperion scanner", "starting scan");
-            NetworkScanner networkScanner = new NetworkScanner();
-
-            String result;
-            while (networkScanner.hasNextAttempt()){
-                result = networkScanner.tryNext();
-
-                if (result != null){
-                    return result;
-                }
-
-                publishProgress(networkScanner.getProgress());
-            }
-
-            return null;
-        }
-
-        @Override
-        protected void onProgressUpdate(Float... values) {
-            Log.d("Hyperion scanner", "scan progress: " + values[0]);
-
-        }
-
-        @Override
-        protected void onPostExecute(String result) {
-            Log.d("Hyperion scanner", "scan result: " + result);
-        }
-    }
 
 
 }
